+++ src/map/utils/charutils.h
@@ -155,7 +155,7 @@ namespace charutils
     void	SaveMissionsList(CCharEntity* PChar);                       // Save the missions list
     void    SaveEminenceData(CCharEntity* PChar);                       // Save Eminence Record (RoE) data
     void    UpdateMissionStorage(CCharEntity* PChar, bool recovery=false); // calculate the rewards for MW1/MW2/MW3/MW4/Satchel and send player a msg if went up
-    bool    HasCompletedMission(CCharEntity* PChar, uint8 log, uint8 mission); // used with UpdateMissionStorage
+    bool    HasCompletedMission(CCharEntity* PChar, uint8 log, uint16 mission); // used with UpdateMissionStorage
     void	SaveQuestsList(CCharEntity* PChar);					        // сохраняем список ксевтов
     void    SaveFame(CCharEntity* PChar);                               // Save area fame / reputation
     void	SaveZonesVisited(CCharEntity* PChar);				        // сохраняем посещенные зоны
